PoliTech/requirements.txt
astroid==2.2.5
colorama==0.4.1
cycler==0.10.0
Django==2.2
djangorestframework==3.9.2
djangorestframework-gis==0.14
dj-database-url==0.5.0
isort==4.3.16
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
mysqlclient==1.4.2
numpy==1.16.2
pip==19.0.3
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
setuptools==40.8.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
wrapt==1.11.1
djangorestframework-bulk==0.2.1
httpie==1.0.2
coreapi==2.3.3
PyYAML==5.1
certifi==2019.3.9
chardet==3.0.4
coreschema==0.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pygments==2.3.1
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.2
